Run trial simulations

For fine control over trial simulations, you must first call run_flfll() and then call either run_engine() or one of the specific engine functions (such as run_engine_contin()). The engines read the *.param files generated by run_flfll(), run the trial simulations, and save output to a bunch of CSV files. You can find these CSV output files next to the *.param files.

run_engine(facts_file, ...)

Arguments

facts_file
Character, name of a FACTS file. Usually has a *.facts file extension.

...
Named arguments to the appropriate inner engine function, such as run_engine_contin(). Use get_facts_engine() to identify the appropriate engine function for your FACTS file. Then, open the help file of that function to read about the arguments.

Value

Nothing.

Details

If you need to repeatedly invoke an engine, as with most trial execution mode workflows, run_engine() is slow. Instead, consider running prep_param_files() and then passing the result to one of the individual engine functions (such as run_engine_contin()).

See also

Examples

# Can only run if system dependencies are configured:
if (file.exists(Sys.getenv("RFACTS_PATHS"))) {
facts_file <- get_facts_file_example("contin.facts") # example FACTS file
out <- run_flfll(facts_file, verbose = FALSE) # Generate param files.
# Run the simulations.
run_engine(
  facts_file,
  param_files = out,
  n_sims = 1,
  verbose = FALSE
)
read_patients(out)
}