Generate param files to prepare for trial simulations

Generate the preparatory files required for simulation.

run_flfll(
  facts_file,
  output_path = tempfile(),
  log_path = output_path,
  n_burn = NULL,
  n_mcmc = NULL,
  n_weeks_files = 10000,
  n_patients_files = 10000,
  n_mcmc_files = 0,
  n_mcmc_thin = NULL,
  flfll_seed = NULL,
  flfll_offset = NULL,
  verbose = FALSE,
  max_sims = 99999L
)

Arguments

facts_file

Character, name of a FACTS file. Usually has a *.facts file extension.

output_path

Character, directory path to the files to generate.

log_path

Character, path to the log file generated by FLFLL.

n_burn

Number of burn-in iterations for the MCMC.

n_mcmc

Number of MCMC iterations used in inference.

n_weeks_files

Number of weeks*.csv files to save in output_path.

n_patients_files

Number of patients*.csv files to save in output_path.

n_mcmc_files

Number of mcmc*.csv files to save in output_path.

n_mcmc_thin

Number of thinning iterations for the MCMC.

flfll_seed

Positive integer, random number generator seed for FLFLL. This seed is only used for stochastic preprocessing steps for generating the *.param files. It is not the random number generator seed for the actual trial simulations. To set the trial simulation seed, use the seed argument of run_facts(), run_engine(), or one of the specific run_engine*() functions.

flfll_offset

Integer, offset for the random number generator.

verbose

Logical, whether to print progress information to the R console.

max_sims

Positive integer of length 1, maximum number of simulations that will be allowed to run for certain engines like CRM in subsequent calls to the engine. If the n_sims argument of the engine is larger than max_sims, only max_sims simulations will be run. The max_sims argument only applies to FLFLL >= 6.4.1 and only needs to be set manually if you are manually calling run_flfll() and then the engine instead of just run_facts().

Value

Character, the value of output_path. output_path is the directory path to the files generated by run_flfll().

Details

For advanced control over trial simulations, you must first call run_flfll() and then call one of the engine functions such as run_engine_contin(). run_flfll() generates the preparatory *.param files that the run_engine_*() functions understand. You will pass these *.param files or their parent directory to param_files argument of run_engine_contin() etc.

See also

Examples

# Can only run if system dependencies are configured:
if (file.exists(Sys.getenv("RFACTS_PATHS"))) {
facts_file <- get_facts_file_example("contin.facts") # example FACTS file
out <- run_flfll(facts_file, verbose = FALSE) # Generate param files.
# Run the simulations.
run_engine(
  facts_file,
  param_files = out,
  n_sims = 1,
  verbose = FALSE
)
read_patients(out)
}